refactoring add autospec tests enable flake8
