+++ .vscode/settings.json
@@ -7,5 +7,6 @@
     "python.testing.unittestEnabled": false,
     "python.testing.pytestEnabled": true,
     "python.analysis.autoImportCompletions": true,
-    "python.analysis.indexing": true
+    "python.analysis.indexing": true,
+    "python.linting.flake8Enabled": true
 }
